# backend/dental/openApi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "description" : "This is the api for the dental module",
    "title" : "Dental Api",
    "version" : "0.1"
  },
  "servers" : [ {
    "url" : "http://localhost:8098"
  } ],
  "paths" : {
...
    "/archiving/procedures/{procedureId}" : {
      "get" : {
        "operationId" : "downloadArchivableProcedure",
        "parameters" : [ {
          "in" : "path",
          "name" : "procedureId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/octet-stream" : { }
            },
            "description" : "OK"
          }
        },
        "summary" : "Get archivable procedure",
        "tags" : [ "Archiving" ]
      }
    },
...
    "/procedures" : {
      "get" : {
        "description" : "GET operation for retrieving basic information of procedures for the procedure overview.\n",
        "operationId" : "getProcedures",
        "parameters" : [ {
          "description" : "Only procedures are returned where the assigneeId of at least one task equals the `assignedToId`.\nIf not submitted, no filtering takes place.\n",
          "in" : "query",
          "name" : "assignedToId",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "description" : "Filter logic:\n- If `procedureType` is submitted, only procedures are returned which have one of the submitted types.\n- If not submitted, no filtering takes place\n",
          "in" : "query",
          "name" : "procedureType",
          "required" : false,
          "schema" : {
            "uniqueItems" : true,
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ProcedureType"
            }
          }
        }, {
          "description" : "Filter logic:\n- If `procedureStatus` is submitted, only procedures are returned which have one of the submitted statuses.\n- If not submitted, no filtering takes place\n",
          "in" : "query",
          "name" : "procedureStatus",
          "required" : false,
          "schema" : {
            "uniqueItems" : true,
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ProcedureStatus"
            }
          }
        }, {
          "description" : "The following sorting options are available:\n- `CREATED_AT`: Sorting by createdAt attribute\n- `MODIFIED_AT`: Sorting by modifiedAt attribute\n",
          "in" : "query",
          "name" : "sortBy",
          "required" : false,
          "schema" : {
            "$ref" : "#/components/schemas/GetProceduresSortBy"
          }
        }, {
          "description" : "Sorting order.",
          "in" : "query",
          "name" : "sortOrder",
          "required" : false,
          "schema" : {
            "$ref" : "#/components/schemas/GetProceduresSortOrder"
          }
        }, {
          "description" : "Limit of returned procedures",
          "in" : "query",
          "name" : "pageSize",
          "required" : false,
          "schema" : {
            "maximum" : 200,
            "minimum" : 1,
            "type" : "integer",
            "format" : "int32",
            "default" : 50
          }
        }, {
          "description" : "Offset used for pagination",
          "in" : "query",
          "name" : "pageNumber",
          "required" : false,
          "schema" : {
            "maximum" : 2000,
            "minimum" : 0,
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GetProceduresResponse"
                }
              }
            },
            "description" : "the users recent procedures"
          }
        },
        "summary" : "Get recent procedures for user",
        "tags" : [ "Procedure" ]
      }
    },
...
    "/procedures/{id}" : {
      "get" : {
        "operationId" : "getDetailedProcedure",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GetDetailedProcedureResponse"
                }
              }
            },
            "description" : "a single procedure with details"
          }
        },
        "summary" : "Get a single procedure with details",
        "tags" : [ "Procedure" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Address" : {
        "required" : [ "@type" ],
        "type" : "object",
        "properties" : {
          "@type" : {
            "type" : "string"
          }
        },
        "discriminator" : {
          "propertyName" : "@type"
        }
      },
...
      "ArchivingRelevance" : {
        "type" : "string",
        "enum" : [ "DEFAULT", "RELEVANT", "IRRELEVANT" ]
      },
      "ArchivingRelevanceSettings" : {
        "required" : [ "archivingRelevance", "defaultArchivingRelevance" ],
        "type" : "object",
        "properties" : {
          "archivingRelevance" : {
            "$ref" : "#/components/schemas/ArchivingRelevance"
          },
          "defaultArchivingRelevance" : {
            "$ref" : "#/components/schemas/ArchivingRelevance"
          }
        }
      },
...
      "BusinessModule" : {
        "type" : "string",
        "enum" : [ "INSPECTION", "SCHOOL_ENTRY", "TRAVEL_MEDICINE", "MEASLES_PROTECTION", "STI_PROTECTION", "MEDICAL_REGISTRY", "DENTAL", "OFFICIAL_MEDICAL_SERVICE", "MEDS_ABROAD" ]
      },
...
      "DataOrigin" : {
        "type" : "string",
        "description" : "A list of possible origins of Persons and Facility in the Central Files. EDIT will only be set automatically on changes. EXTERNAL is for entries that come, e.g., from the citizen portal. IMPORT is reserved for automatic imports. MANUAL shall be set for every creation or connection done by an employee.",
        "enum" : [ "MANUAL", "EXTERNAL", "IMPORT", "EDIT" ]
      },
...
      "DetailedFacility" : {
        "required" : [ "facilityFileState", "facilityType" ],
        "type" : "object",
        "properties" : {
          "facilityFileState" : {
            "$ref" : "#/components/schemas/GetFacilityFileStateResponse"
          },
          "facilityType" : {
            "$ref" : "#/components/schemas/FacilityType"
          }
        }
      },
      "DetailedPerson" : {
        "required" : [ "person", "personType" ],
        "type" : "object",
        "properties" : {
          "person" : {
            "$ref" : "#/components/schemas/GetPersonFileStateResponse"
          },
          "personType" : {
            "$ref" : "#/components/schemas/PersonType"
          }
        }
      },
      "DetailedTask" : {
        "required" : [ "task" ],
        "type" : "object",
        "properties" : {
          "assignedByName" : {
            "type" : "string"
          },
          "assigneeName" : {
            "type" : "string"
          },
          "task" : {
            "$ref" : "#/components/schemas/Task"
          }
        }
      },
...
      "DomesticAddress" : {
        "required" : [ "city", "country", "postalCode", "street" ],
        "type" : "object",
        "description" : "A usual domestic address.",
        "allOf" : [ {
          "$ref" : "#/components/schemas/Address"
        }, {
          "type" : "object",
          "properties" : {
            "addressAddition" : {
              "maxLength" : 100,
              "minLength" : 1,
              "type" : "string",
              "description" : "A descriptive addition to the address.",
              "example" : "2.OG links"
            },
            "city" : {
              "maxLength" : 50,
              "minLength" : 1,
              "type" : "string",
              "description" : "The city in which the address is located.",
              "example" : "Berlin"
            },
            "country" : {
              "$ref" : "#/components/schemas/CountryCode"
            },
            "differentName" : {
              "maxLength" : 200,
              "minLength" : 1,
              "type" : "string",
              "description" : "If the name of the addressee deviates from the present name of a Person or Facility, e.g. when a company pays the bill for its employee or a parent company shall be contacted instead of its subsidiary.",
              "example" : "Parent Company AG"
            },
            "houseNumber" : {
              "maxLength" : 11,
              "minLength" : 1,
              "type" : "string",
              "description" : "The house number of the address, including extensions.",
              "example" : "1a"
            },
            "postalCode" : {
              "maxLength" : 20,
              "minLength" : 1,
              "type" : "string",
              "description" : "The postal code of the address.",
              "example" : "10115"
            },
            "street" : {
              "maxLength" : 55,
              "minLength" : 1,
              "type" : "string",
              "description" : "The name of the street of the address, without the house number.",
              "example" : "Beispielweg"
            }
          }
        } ]
      },
...
      "FacilityContactPerson" : {
        "required" : [ "lastName" ],
        "type" : "object",
        "properties" : {
          "emailAddress" : {
            "maxLength" : 254,
            "minLength" : 6,
            "type" : "string",
            "description" : "The email addresses of the Contact Person.",
            "example" : "mail1@address.de"
          },
          "firstName" : {
            "maxLength" : 80,
            "minLength" : 1,
            "type" : "string",
            "description" : "The given name(s) of the Contact Person.",
            "example" : "John"
          },
          "gender" : {
            "$ref" : "#/components/schemas/Gender"
          },
          "lastName" : {
            "maxLength" : 120,
            "minLength" : 1,
            "type" : "string",
            "description" : "The last name of the Contact Person.",
            "example" : "Doe"
          },
          "mainContact" : {
            "type" : "boolean"
          },
          "phoneNumber" : {
            "maxLength" : 23,
            "minLength" : 1,
            "type" : "string",
            "description" : "The phone number of the Contact Person.",
            "example" : "+491234567890"
          },
          "role" : {
            "maxLength" : 255,
            "minLength" : 1,
            "type" : "string",
            "description" : "The role of the Contact Person in the Facility.",
            "example" : "CEO"
          },
          "salutation" : {
            "$ref" : "#/components/schemas/Salutation"
          },
          "title" : {
            "maxLength" : 119,
            "minLength" : 1,
            "type" : "string",
            "description" : "The academic title of the Contact Person.",
            "example" : "Prof. Dr."
          }
        }
      },
      "FacilityType" : {
        "type" : "string",
        "enum" : [ "SCHOOL", "INSPECTION", "DAYCARE", "HOSPITAL", "MEDICAL_PRACTICE", "REFUGEE_ACCOMMODATION", "OTHER" ]
      },
...
      "Gender" : {
        "type" : "string",
        "description" : "The list of genders as specified in the German Personenstandsgesetz.",
        "enum" : [ "NOT_SPECIFIED", "DIVERSE", "FEMALE", "MALE" ]
      },
...
      "GetDetailedProcedureResponse" : {
        "required" : [ "facilities", "persons", "procedure", "tasks" ],
        "type" : "object",
        "properties" : {
          "facilities" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/DetailedFacility"
            }
          },
          "persons" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/DetailedPerson"
            }
          },
          "procedure" : {
            "$ref" : "#/components/schemas/Procedure"
          },
          "tasks" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/DetailedTask"
            }
          }
        }
      },
      "GetFacilityFileStateResponse" : {
        "required" : [ "contactPersons", "dataOrigin", "emailAddresses", "id", "name", "phoneNumbers", "referenceVersion" ],
        "type" : "object",
        "properties" : {
          "contactAddress" : {
            "oneOf" : [ {
              "$ref" : "#/components/schemas/DomesticAddress"
            }, {
              "$ref" : "#/components/schemas/PostboxAddress"
            } ]
          },
          "contactPersons" : {
            "type" : "array",
            "description" : "A list of contact persons of the Facility.",
            "items" : {
              "$ref" : "#/components/schemas/FacilityContactPerson"
            }
          },
          "dataOrigin" : {
            "$ref" : "#/components/schemas/DataOrigin"
          },
          "differentBillingAddress" : {
            "oneOf" : [ {
              "$ref" : "#/components/schemas/DomesticAddress"
            }, {
              "$ref" : "#/components/schemas/PostboxAddress"
            } ]
          },
          "emailAddresses" : {
            "type" : "array",
            "description" : "A list of email addresses of the Facility.",
            "example" : "['mail1@address.de','mail2@address.de','mail3@address.de']",
            "items" : {
              "type" : "string"
            }
          },
          "id" : {
            "type" : "string",
            "description" : "Id of the Facility.",
            "format" : "uuid",
            "example" : "UUID_1"
          },
          "name" : {
            "maxLength" : 300,
            "minLength" : 1,
            "type" : "string",
            "description" : "The name of the Facility.",
            "example" : "Example Facility"
          },
          "outdated" : {
            "type" : "boolean",
            "description" : "A flag that signals if a File State differs from the referenceFacility it is connected to.",
            "example" : true
          },
          "phoneNumbers" : {
            "type" : "array",
            "description" : "A list of telephone numbers of the Facility.",
            "example" : "['+4912345678901','+4912345678902','+4912345678903']",
            "items" : {
              "maxLength" : 23,
              "minLength" : 1,
              "type" : "string"
            }
          },
          "referenceVersion" : {
            "type" : "integer",
            "description" : "The version of referenceData that was present when the FileState was created. Can be increased if a newer version is irrelevant for the Procedure and the outdated flag shall be suppressed.",
            "format" : "int64",
            "example" : 1
          }
        }
      },
...
      "GetPersonFileStateResponse" : {
        "required" : [ "dataOrigin", "dateOfBirth", "emailAddresses", "firstName", "gender", "humanReadableId", "id", "lastName", "phoneNumbers", "referenceVersion", "salutation" ],
        "type" : "object",
        "properties" : {
          "contactAddress" : {
            "oneOf" : [ {
              "$ref" : "#/components/schemas/DomesticAddress"
            }, {
              "$ref" : "#/components/schemas/PostboxAddress"
            } ]
          },
          "countryOfBirth" : {
            "$ref" : "#/components/schemas/CountryCode"
          },
          "dataOrigin" : {
            "$ref" : "#/components/schemas/DataOrigin"
          },
          "dateOfBirth" : {
            "type" : "string",
            "description" : "The date of birth of the Person.",
            "format" : "date",
            "example" : "2000-01-01"
          },
          "differentBillingAddress" : {
            "oneOf" : [ {
              "$ref" : "#/components/schemas/DomesticAddress"
            }, {
              "$ref" : "#/components/schemas/PostboxAddress"
            } ]
          },
          "emailAddresses" : {
            "type" : "array",
            "description" : "A list of email addresses of the Person.",
            "example" : "['mail1@address.de','mail2@address.de','mail3@address.de']",
            "items" : {
              "type" : "string"
            }
          },
          "firstName" : {
            "maxLength" : 80,
            "minLength" : 1,
            "type" : "string",
            "description" : "The given name(s) of the Person.",
            "example" : "John"
          },
          "gender" : {
            "$ref" : "#/components/schemas/Gender"
          },
          "humanReadableId" : {
            "maxLength" : 8,
            "minLength" : 8,
            "type" : "string"
          },
          "id" : {
            "type" : "string",
            "description" : "Id of the Person.",
            "format" : "uuid",
            "example" : "UUID_1"
          },
          "lastName" : {
            "maxLength" : 120,
            "minLength" : 1,
            "type" : "string",
            "description" : "The last name of the Person.",
            "example" : "Doe"
          },
          "nameAtBirth" : {
            "maxLength" : 40,
            "minLength" : 1,
            "type" : "string",
            "description" : "The last name at birth of the Person.",
            "example" : "Smith"
          },
          "outdated" : {
            "type" : "boolean",
            "description" : "A flag that signals if a File State differs from the referencePerson it is connected to. Set to 'null' if unchecked.",
            "example" : true
          },
          "phoneNumbers" : {
            "type" : "array",
            "description" : "A list of telephone numbers of the Person.",
            "example" : "['+4912345678901','+4912345678902','+4912345678903']",
            "items" : {
              "maxLength" : 23,
              "minLength" : 1,
              "type" : "string"
            }
          },
          "placeOfBirth" : {
            "maxLength" : 50,
            "minLength" : 1,
            "type" : "string",
            "description" : "The place of birth (without country) of the Person.",
            "example" : "Berlin"
          },
          "referenceVersion" : {
            "type" : "integer",
            "description" : "The version of referenceData that was present when the FileState was created. Can be increased if a newer version is irrelevant for the Procedure and the outdated flag shall be suppressed.",
            "format" : "int64",
            "example" : 1
          },
          "salutation" : {
            "$ref" : "#/components/schemas/Salutation"
          },
          "title" : {
            "maxLength" : 119,
            "minLength" : 1,
            "type" : "string",
            "description" : "The academic title of a Person.",
            "example" : "Prof. Dr."
          }
        }
      },
...
      "GetProceduresResponse" : {
        "required" : [ "procedures", "totalElements", "totalPages" ],
        "type" : "object",
        "properties" : {
          "procedures" : {
            "maxItems" : 200,
            "minItems" : 0,
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Procedure"
            }
          },
          "totalElements" : {
            "type" : "integer",
            "description" : "Total number of result elements for the given filter criteria",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "description" : "Total number of result pages for the given filter criteria",
            "format" : "int32"
          }
        }
      },
      "GetProceduresSortBy" : {
        "type" : "string",
        "default" : "CREATED_AT",
        "enum" : [ "MODIFIED_AT", "CREATED_AT" ]
      },
      "GetProceduresSortOrder" : {
        "type" : "string",
        "default" : "ASC",
        "enum" : [ "ASC", "DESC" ]
      },
...
      "PersonType" : {
        "type" : "string",
        "enum" : [ "PATIENT", "PARENT", "PROFESSIONAL", "EMPLOYEE" ]
      },
...
      "PostboxAddress" : {
        "required" : [ "city", "country", "postalCode", "postbox" ],
        "type" : "object",
        "description" : "An address which is a postbox.",
        "allOf" : [ {
          "$ref" : "#/components/schemas/Address"
        }, {
          "type" : "object",
          "properties" : {
            "city" : {
              "maxLength" : 50,
              "minLength" : 1,
              "type" : "string",
              "description" : "The city in which the address is located.",
              "example" : "Berlin"
            },
            "country" : {
              "$ref" : "#/components/schemas/CountryCode"
            },
            "differentName" : {
              "maxLength" : 200,
              "minLength" : 1,
              "type" : "string",
              "description" : "If the name of the addressee deviates from the present name of a Person or Facility, e.g. when a company pays the bill for its employee or a parent company shall be contacted instead of its subsidiary.",
              "example" : "Parent Company AG"
            },
            "postalCode" : {
              "maxLength" : 20,
              "minLength" : 1,
              "type" : "string",
              "description" : "The postal code of the address.",
              "example" : "10115"
            },
            "postbox" : {
              "maxLength" : 21,
              "minLength" : 1,
              "type" : "string",
              "description" : "The number (or name) of the postbox.",
              "example" : "123"
            }
          }
        } ]
      },
      "Procedure" : {
        "required" : [ "archivingRelevanceSettings", "businessModule", "createdAt", "modifiedAt", "procedureId", "procedureStatus", "procedureType", "summary" ],
        "type" : "object",
        "properties" : {
          "archivingRelevanceSettings" : {
            "$ref" : "#/components/schemas/ArchivingRelevanceSettings"
          },
          "businessModule" : {
            "$ref" : "#/components/schemas/BusinessModule"
          },
          "closedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "exportedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "modifiedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "procedureId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "procedureStatus" : {
            "$ref" : "#/components/schemas/ProcedureStatus"
          },
          "procedureType" : {
            "$ref" : "#/components/schemas/ProcedureType"
          },
          "summary" : {
            "maxLength" : 128,
            "minLength" : 0,
            "pattern" : "[a-zA-Z0-9.].+",
            "type" : "string"
          }
        }
      },
...
      "ProcedureStatus" : {
        "type" : "string",
        "enum" : [ "DRAFT", "OPEN", "IN_PROGRESS", "CLOSED", "ABORTED" ]
      },
      "ProcedureType" : {
        "type" : "string",
        "enum" : [ "REGULAR_EXAMINATION", "CAN_CHILD", "ENTRY_LEVEL", "DRAFT_CITIZEN_OFFICE_IMPORT", "DRAFT_SCHOOL_IMPORT", "INSPECTION", "TM_VACCINATION_CONSULTATION", "MEASLES_PROTECTION", "STI_PROTECTION", "MEDICAL_REGISTRY_ENTRY", "MEDICAL_REGISTRY_CITIZEN_DRAFT", "MEDICAL_REGISTRY_EMPLOYEE_DRAFT", "DENTAL_CHILD", "OFFICIAL_MEDICAL_SERVICE", "MEDS_ABROAD" ]
      },
...
      "Salutation" : {
        "type" : "string",
        "enum" : [ "NOT_SPECIFIED", "NEUTRAL", "FEMALE", "MALE" ]
      },
...
      "Task" : {
        "required" : [ "businessModule", "createdAt", "isOverdue", "modifiedAt", "procedureId", "summary", "taskId", "taskStatus", "taskType", "version" ],
        "type" : "object",
        "properties" : {
          "assignedById" : {
            "type" : "string",
            "format" : "uuid"
          },
          "assigneeId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "businessModule" : {
            "$ref" : "#/components/schemas/BusinessModule"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "dueAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "isOverdue" : {
            "type" : "boolean"
          },
          "modifiedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "procedureId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "summary" : {
            "maxLength" : 128,
            "minLength" : 0,
            "type" : "string"
          },
          "taskId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "taskStatus" : {
            "$ref" : "#/components/schemas/TaskStatus"
          },
          "taskType" : {
            "$ref" : "#/components/schemas/TaskType"
          },
          "version" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
      "TaskStatus" : {
        "type" : "string",
        "enum" : [ "OPEN", "CLOSED" ]
      },
      "TaskType" : {
        "type" : "string",
        "enum" : [ "BOOK_APPOINTMENT", "PERFORM_SCHOOL_ENTRY_EXAMINATION", "INSPECTION_PLANNING", "INSPECTION_EXECUTION", "INSPECTION_REPORT", "TRAVEL_MEDICINE", "MEASLES_PROTECTION", "STI_PROTECTION", "OFFICIAL_MEDICAL_SERVICE", "MEDS_ABROAD" ]
      },
...
    }
  }
}